Sprint 1 - Igho
AFter successfully working through lab 8 (Inserting date data into the dates table in destinationDatabase.mdf), i have now applied the same code to the products data that was pulled out of the source (sheet1) with the intent of inserting the data into the products table of the mdf file. The code builds and runs successfully without any errors, but when i viewed the table data of the auto-generated destination db(the debug one), there were no entries inserted. I tried to use the console.writeline debug tool to no avail. The coding task for this week was dedicated to Zhaofeng because it was my turn last week but i have gone ahead to try on my own for the benefit of experience. Zhaofeng has also done the coding as required, his code runs, but 'catches an exception', i tried to re-arrange the order of the 'Insert' statement to match with the order of the'Insert Parameter lines', but the code still breaks. However, when i checked the auto-generated db, there were a few entries. I have looked through his code. everything checks out with mine so i suggested we both inspect each others branch for any notable differences as to why i cannot successfully insert data. Zhaofeng spotted that i omitted the splitProduct function were i was meant to assign variables to their respective array positions, and also did not call the splitFunction in the GetProduct eventhandler. Even after adjusting accordingly, i still cannot insert data in my own version of the code. We will continue trying to fix this until our next lab.